DiscordSnapshotService: Start working on logic for reducing spammy member snapshot events

+++ XeniaDiscord.Data/Models/Snapshot/GuildMemberSnapshotEventModel.cs
@@ -73,7 +73,7 @@ public class GuildMemberSnapshotEventModel
 }
 
 [Flags]
-public enum GuildMemberSnapshotEventWhatChanged : ulong
+public enum GuildMemberSnapshotEventWhatChanged : byte
 {
     /// <summary>
     /// <see cref="GuildMemberSnapshotModel.Username"/>
